fix: - Code optimized - Tests divided into valid and invalid values - Removed comments
